+++ src/ra/mapeddlg.cc
@@ -1464,7 +1464,7 @@ int MapEditClass::Load_Scenario() {
         *WWLIB doesn't pass through a KN_MOUSE_MOVE *	  value while the button
         *is being held down, so this case must be *	  trapped as a default.
         */
-        if (static_cast<int>(input) == ((int)KN_LMOUSE | (int)KN_RLSE_BIT)) {
+        if (static_cast<int>(input) == (KN_LMOUSE | KN_RLSE_BIT)) {
           grabbed = 0;
           display = REDRAW_MAP;
         } else {
