+++ src/td/mapedit.cc
@@ -849,7 +849,7 @@ void MapEditClass::AI(KeyNumType& input, int x, int y) {
     /*---------------------------------------------------------------------
     SHIFT-HOME: set new Home Cell position
     ---------------------------------------------------------------------*/
-    case ((int)KN_HOME | (int)KN_SHIFT_BIT):
+    case (KN_HOME | KN_SHIFT_BIT):
       /*
       ** Unflag the old Home Cell, if there are no other waypoints
       ** pointing to it
@@ -884,7 +884,7 @@ void MapEditClass::AI(KeyNumType& input, int x, int y) {
     SHIFT-R: set new Reinforcement Cell position.  Don't allow setting
     the Reinf. Cell to the same as the Home Cell (for display purposes.)
     ---------------------------------------------------------------------*/
-    case ((int)KN_R | (int)KN_SHIFT_BIT):
+    case (KN_R | KN_SHIFT_BIT):
       if (CurrentCell == 0 || CurrentCell == Waypoint[WAYPT_HOME]) {
         break;
       }
@@ -921,32 +921,32 @@ void MapEditClass::AI(KeyNumType& input, int x, int y) {
     /*---------------------------------------------------------------------
     ALT-Letter: Label a waypoint cell
     ---------------------------------------------------------------------*/
-    case ((int)KN_A | (int)KN_ALT_BIT):
-    case ((int)KN_B | (int)KN_ALT_BIT):
-    case ((int)KN_C | (int)KN_ALT_BIT):
-    case ((int)KN_D | (int)KN_ALT_BIT):
-    case ((int)KN_E | (int)KN_ALT_BIT):
-    case ((int)KN_F | (int)KN_ALT_BIT):
-    case ((int)KN_G | (int)KN_ALT_BIT):
-    case ((int)KN_H | (int)KN_ALT_BIT):
-    case ((int)KN_I | (int)KN_ALT_BIT):
-    case ((int)KN_J | (int)KN_ALT_BIT):
-    case ((int)KN_K | (int)KN_ALT_BIT):
-    case ((int)KN_L | (int)KN_ALT_BIT):
-    case ((int)KN_M | (int)KN_ALT_BIT):
-    case ((int)KN_N | (int)KN_ALT_BIT):
-    case ((int)KN_O | (int)KN_ALT_BIT):
-    case ((int)KN_P | (int)KN_ALT_BIT):
-    case ((int)KN_Q | (int)KN_ALT_BIT):
-    case ((int)KN_R | (int)KN_ALT_BIT):
-    case ((int)KN_S | (int)KN_ALT_BIT):
-    case ((int)KN_T | (int)KN_ALT_BIT):
-    case ((int)KN_U | (int)KN_ALT_BIT):
-    case ((int)KN_V | (int)KN_ALT_BIT):
-    case ((int)KN_W | (int)KN_ALT_BIT):
-    case ((int)KN_X | (int)KN_ALT_BIT):
-    case ((int)KN_Y | (int)KN_ALT_BIT):
-    case ((int)KN_Z | (int)KN_ALT_BIT):
+    case (KN_A | KN_ALT_BIT):
+    case (KN_B | KN_ALT_BIT):
+    case (KN_C | KN_ALT_BIT):
+    case (KN_D | KN_ALT_BIT):
+    case (KN_E | KN_ALT_BIT):
+    case (KN_F | KN_ALT_BIT):
+    case (KN_G | KN_ALT_BIT):
+    case (KN_H | KN_ALT_BIT):
+    case (KN_I | KN_ALT_BIT):
+    case (KN_J | KN_ALT_BIT):
+    case (KN_K | KN_ALT_BIT):
+    case (KN_L | KN_ALT_BIT):
+    case (KN_M | KN_ALT_BIT):
+    case (KN_N | KN_ALT_BIT):
+    case (KN_O | KN_ALT_BIT):
+    case (KN_P | KN_ALT_BIT):
+    case (KN_Q | KN_ALT_BIT):
+    case (KN_R | KN_ALT_BIT):
+    case (KN_S | KN_ALT_BIT):
+    case (KN_T | KN_ALT_BIT):
+    case (KN_U | KN_ALT_BIT):
+    case (KN_V | KN_ALT_BIT):
+    case (KN_W | KN_ALT_BIT):
+    case (KN_X | KN_ALT_BIT):
+    case (KN_Y | KN_ALT_BIT):
+    case (KN_Z | KN_ALT_BIT):
       if (CurrentCell != 0) {
         waypt_idx = KN_To_KA(input & 0xff) - KA_a;
         /*...............................................................
@@ -970,10 +970,10 @@ void MapEditClass::AI(KeyNumType& input, int x, int y) {
     /*---------------------------------------------------------------------
     ALT-1-4: Designate a cell as a capture-the-flag cell.
     ---------------------------------------------------------------------*/
-    case ((int)KN_1 | (int)KN_ALT_BIT):
-    case ((int)KN_2 | (int)KN_ALT_BIT):
-    case ((int)KN_3 | (int)KN_ALT_BIT):
-    case ((int)KN_4 | (int)KN_ALT_BIT):
+    case (KN_1 | KN_ALT_BIT):
+    case (KN_2 | KN_ALT_BIT):
+    case (KN_3 | KN_ALT_BIT):
+    case (KN_4 | KN_ALT_BIT):
       /*------------------------------------------------------------------
       If there's a current cell, place the flag & waypoint there.
       ------------------------------------------------------------------*/
@@ -1004,7 +1004,7 @@ void MapEditClass::AI(KeyNumType& input, int x, int y) {
     /*---------------------------------------------------------------------
     ALT-Space: Remove a waypoint designation
     ---------------------------------------------------------------------*/
-    case ((int)KN_SPACE | (int)KN_ALT_BIT):
+    case (KN_SPACE | KN_ALT_BIT):
       if (CurrentCell != 0) {
         /*...............................................................
         Loop through letter waypoints; if this cell is one of them,
@@ -1047,7 +1047,7 @@ void MapEditClass::AI(KeyNumType& input, int x, int y) {
     'H' = toggle current placement object's house
     ---------------------------------------------------------------------*/
     case KN_H:
-    case ((int)KN_H | (int)KN_SHIFT_BIT):
+    case (KN_H | KN_SHIFT_BIT):
       if (PendingObject) {
         Toggle_House();
       }
@@ -1129,10 +1129,10 @@ void MapEditClass::AI(KeyNumType& input, int x, int y) {
     /*---------------------------------------------------------------------
     SHIFT-ALT-Arrow: move the current object
     ---------------------------------------------------------------------*/
-    case (int)KN_UP | (int)KN_ALT_BIT | (int)KN_SHIFT_BIT:
-    case (int)KN_DOWN | (int)KN_ALT_BIT | (int)KN_SHIFT_BIT:
-    case (int)KN_LEFT | (int)KN_ALT_BIT | (int)KN_SHIFT_BIT:
-    case (int)KN_RIGHT | (int)KN_ALT_BIT | (int)KN_SHIFT_BIT:
+    case KN_UP | KN_ALT_BIT | KN_SHIFT_BIT:
+    case KN_DOWN | KN_ALT_BIT | KN_SHIFT_BIT:
+    case KN_LEFT | KN_ALT_BIT | KN_SHIFT_BIT:
+    case KN_RIGHT | KN_ALT_BIT | KN_SHIFT_BIT:
       if (CurrentObject.Count()) {
         CurrentObject[0]->Move(KN_To_Facing(input));
         Changed = true;
